Fixed PhysX build issues
