+++ Code/EnginePlugins/PhysXPlugin/WorldModule/Implementation/PhysXSimulationEvents.cpp
@@ -4,9 +4,10 @@
 #include <Core/Prefabs/PrefabResource.h>
 #include <Core/WorldSerializer/WorldReader.h>
 #include <Foundation/Configuration/CVar.h>
+#include <GameEngine/Physics/SurfaceResource.h>
 #include <PhysXPlugin/Components/PxTriggerComponent.h>
 #include <PhysXPlugin/Utilities/PxConversionUtils.h>
-#include <PhysXPlugin/WorldModule/Implementation/PhysX.h>
+#include <PhysXPlugin/WorldModule/Implementation/PhysXSimulationEvents.h>
 #include <PhysXPlugin/WorldModule/PhysXWorldModule.h>
 #include <RendererCore/Debug/DebugRenderer.h>
 
@@ -16,71 +17,6 @@ ezCVarBool cvar_PhysicsReactionsVisImpacts("Physics.Reactions.VisImpacts", false
 ezCVarBool cvar_PhysicsReactionsVisDiscardedImpacts("Physics.Reactions.VisDiscardedImpacts", false, ezCVarFlags::Default, "Visualize where impact reactions were NOT spawned.");
 ezCVarBool cvar_PhysicsReactionsVisSlides("Physics.Reactions.VisSlides", false, ezCVarFlags::Default, "Visualize active slide reactions.");
 ezCVarBool cvar_PhysicsReactionsVisRolls("Physics.Reactions.VisRolls", false, ezCVarFlags::Default, "Visualize active roll reactions.");
-
-class ezPxSimulationEventCallback : public physx::PxSimulationEventCallback
-{
-public:
-  struct InteractionContact
-  {
-    ezVec3 m_vPosition;
-    ezVec3 m_vNormal;
-    ezSurfaceResource* m_pSurface;
-    ezTempHashedString m_sInteraction;
-    float m_fImpulseSqr;
-    float m_fDistanceSqr;
-  };
-
-  struct SlideAndRollInfo
-  {
-    physx::PxRigidDynamic* m_pActor = nullptr;
-    bool m_bStillSliding = false;
-    bool m_bStillRolling = false;
-
-    float m_fDistanceSqr;
-    ezVec3 m_vContactPosition;
-    ezGameObjectHandle m_hSlidePrefab;
-    ezGameObjectHandle m_hRollPrefab;
-    ezHashedString m_sSlideInteractionPrefab;
-    ezHashedString m_sRollInteractionPrefab;
-  };
-
-  ezVec3 m_vMainCameraPosition = ezVec3::ZeroVector();
-  ezHybridArray<InteractionContact, 8> m_InteractionContacts; // these are spawned PER FRAME, so only a low number is necessary
-  ezHybridArray<SlideAndRollInfo, 4> m_SlidingOrRollingActors;
-  ezDeque<physx::PxConstraint*> m_BrokenConstraints;
-  ezWorld* m_pWorld = nullptr;
-
-  SlideAndRollInfo* FindSlideOrRollInfo(physx::PxRigidDynamic* pActor, const ezVec3& vAvgPos);
-
-  virtual void onConstraintBreak(physx::PxConstraintInfo* constraints, physx::PxU32 count) override;
-
-  virtual void onContact(const physx::PxContactPairHeader& pairHeader, const physx::PxContactPair* pairs, physx::PxU32 nbPairs) override;
-
-  void OnContact_SlideReaction(const ezVec3& vAvgPos, const ezVec3& vAvgNormal, physx::PxRigidDynamic* pRigid0, physx::PxRigidDynamic* pRigid1, const physx::PxContactPair& pair, ezBitflags<ezOnPhysXContact>& ContactFlags0, ezBitflags<ezOnPhysXContact>& ContactFlags1);
-  void OnContact_RollReaction(const ezVec3& vAvgPos, physx::PxRigidDynamic* pRigid0, physx::PxRigidDynamic* pRigid1, const physx::PxContactPair& pair, ezBitflags<ezOnPhysXContact>& ContactFlags0, ezBitflags<ezOnPhysXContact>& ContactFlags1);
-  void OnContact_SlideAndRollReaction(const physx::PxContactPairHeader& pairHeader, const physx::PxContactPair& pair, ezBitflags<ezOnPhysXContact>& ContactFlags0, const ezVec3& vAvgPos, const ezVec3& vAvgNormal, ezBitflags<ezOnPhysXContact>& ContactFlags1, const ezUInt32 uiNumContactPoints, ezBitflags<ezOnPhysXContact>& CombinedContactFlags);
-
-  void OnContact_ImpactReaction(physx::PxContactPairPoint* contactPointBuffer, const physx::PxContactPair& pair, const ezVec3& vAvgPos, const ezVec3& vAvgNormal, float fMaxImpactSqr, const physx::PxContactPairHeader& pairHeader);
-
-  void SendContactReport(const physx::PxContactPairHeader& pairHeader, const physx::PxContactPair* pairs, physx::PxU32 nbPairs);
-
-  struct TriggerEvent
-  {
-    ezComponentHandle m_hTriggerComponent;
-    ezComponentHandle m_hOtherComponent;
-    ezTriggerState::Enum m_TriggerState;
-  };
-
-  ezDeque<TriggerEvent> m_TriggerEvents;
-
-  virtual void onTrigger(physx::PxTriggerPair* pairs, physx::PxU32 count) override;
-
-  virtual void onWake(physx::PxActor** actors, physx::PxU32 count) override {}
-
-  virtual void onSleep(physx::PxActor** actors, physx::PxU32 count) override {}
-
-  virtual void onAdvance(const physx::PxRigidBody* const* bodyBuffer, const physx::PxTransform* poseBuffer, const physx::PxU32 count) override {}
-};
 
 void ezPxSimulationEventCallback::onConstraintBreak(PxConstraintInfo* constraints, PxU32 count)
 {
